[Savestates] Split SavestatePlayerHandler into client and server
